Fixed emissive textures not rendered properly.
In dxvk-remix textures rendered with certain alpha blending modes are interpreted as emissive textures, however, this would also cause transparency in the rendered textures.
The renderer now renders the affected geometry thrice, once without the unlit flag, and once scaled up and once scaled down by a tiny bit (since the point of origin for the mesh might be behind the camera).